// front-end/src/app/core/models/genre.ts
/** Book genres. */
export enum Genre {
  History = 'History',
  Horror = 'Horror',
  Travel = 'Travel',
  Art = 'Art',
  Business = 'Business',
  Mystery = 'Mystery',
  Fantasy = 'Fantasy',
  Thriller = 'Thriller',
  Comics = 'Comics',
  SciFi = 'Science-Fiction',
  NonFiction = 'Non-Fiction',
  Science = 'Science',
  Biography = 'Biography',
}