// front-end/src/app/core/services/dtos/genre-dto.ts
/** Genre Dto. */
export enum GenreDto {
  History = 'history',
  Horror = 'horror',
  Travel = 'travel',
  Art = 'art',
  Business = 'business',
  Mystery = 'mystery',
  Fantasy = 'fantasy',
  Thriller = 'thriller',
  Comics = 'comics',
  SciFi = 'science-fiction',
  NonFiction = 'non-fiction',
  Science = 'science',
  Biography = 'biography',
}